Update lone pairs as first step in molecule.updateAtomType()
When lone pairs are left unspecified for groups, we label it as -100 as a placeholder because when we interface with RDKit, it requires an integer. The new PR requires lone pairs be specified to something sensible before assigning atomtype. Therefore we have to update lone pairs prior to updating atomtype.

Also removed some instances of updateLonePairs() after updateAtomType() as they are now redundant.