Added a saturateH option to generateIsomorphicResonanceStructures
If `saturateH` is `True`, then saturate `mol` with hydrogens before
generating the resonance structures and remove the hydrogens before
returning `isomorphic_isomers`. This is useful when resonance structures
are generated for molecules in which all hydrogens were intentionally
removed as in generating augInChI. Otherwise, RMG will probably get many
of the lonePairs and partial charges in a molecule wrong.