Fix TestThermoAccuracyAromatics
The original species in the test, 'c([O])1c([C]=O)cc(C=O)cc1',
has a singlet resonance structure (with a more stable thermo).
Therefore, the expected comment will not appear in thermo.comment
The fix includes removing the radical from the oxygen site (which
isn't tested here). Since this results in 'p_OH_CHO' being added
to thermo.comment, that atom was changed to nitrogen to keep the
test assertions as is.